Bug Fix! Image size calculated incorrectly.
The imageSize variable in the decode function of the FCode class was being wrongly calculated in some specific cases. The variable was increased by 1 whenever the number of characters allowed a perfect square. That occurred because they were being calculated with the nextPerfectSquareSide() function, even when it was a perfect square. The IsPerfectSquare() function was added to check if the float result from the ceil(textLength/ (float)4) is a perfect square and if so, set the variable to the square root of that value. If not, then the code will use the nextPerfectSquareSide() function.

+new scheme2.png image added.
+scheme2.psd file added.
+scheme.psd file added.